standalone header generation tests
